Various improvements in lyon_geom.
 - Improved CubicBezierSegment::for_each_quadratic_bezier (generates less segments)
 - Faster and more precise CubicBezierSegment::is_linear
 - Fixed Line::distance_to_point
 - Fixed CubicBezierSegment::for_each_monotonic
 - Added CubicBezierSegment::num_quadratics
 - Added CubicBezierSegment::to_quadratic/to_quadratic_error/is_quadratic
 - Moved the code from flatten_cubic.rs and cubic_to_quadratic.rs into cubic_bezier.rs
