+++ crates/geom/src/cubic_bezier.rs
@@ -1,7 +1,4 @@
 use crate::cubic_bezier_intersections::cubic_bezier_intersections_t;
-use crate::cubic_to_quadratic::*;
-pub use crate::flatten_cubic::Flattened;
-use crate::flatten_cubic::{find_cubic_bezier_inflection_points, flatten_cubic_bezier_with_t};
 use crate::scalar::Scalar;
 use crate::segment::{BoundingBox, Segment};
 use crate::traits::Transformation;
@@ -71,7 +68,6 @@ impl<S: Scalar> CubicBezierSegment<S> {
     }
 
     /// Return the parameter values corresponding to a given x coordinate.
-    /// See also solve_t_for_x for monotonic curves.
     pub fn solve_t_for_x(&self, x: S) -> ArrayVec<S, 3> {
         if self.is_a_point(S::ZERO)
             || (self.non_point_is_linear(S::ZERO) && self.from.x == self.to.x)
@@ -83,7 +79,6 @@ impl<S: Scalar> CubicBezierSegment<S> {
     }
 
     /// Return the parameter values corresponding to a given y coordinate.
-    /// See also solve_t_for_y for monotonic curves.
     pub fn solve_t_for_y(&self, y: S) -> ArrayVec<S, 3> {
         if self.is_a_point(S::ZERO)
             || (self.non_point_is_linear(S::ZERO) && self.from.y == self.to.y)
@@ -243,6 +238,8 @@ impl<S: Scalar> CubicBezierSegment<S> {
         }
     }
 
+    /// Returns true if the curve can be approximated with a single line segment, given
+    /// a tolerance threshold.
     pub fn is_linear(&self, tolerance: S) -> bool {
         let epsilon = S::EPSILON;
         if (self.from - self.to).square_length() < epsilon {
@@ -254,9 +251,27 @@ impl<S: Scalar> CubicBezierSegment<S> {
 
     #[inline]
     fn non_point_is_linear(&self, tolerance: S) -> bool {
-        let line = self.baseline().to_line().equation();
-        line.distance_to_point(&self.ctrl1) <= tolerance
-            && line.distance_to_point(&self.ctrl2) <= tolerance
+        // Similar to Line::square_distance_to_point, except we keep
+        // the sign of c1 and c2 to compute tighter upper bounds as we
+        // do in fat_line_min_max.
+        let baseline = self.to - self.from;
+        let v1 = self.ctrl1 - self.from;
+        let v2 = self.ctrl2 - self.from;
+        let c1 = baseline.cross(v1);
+        let c2 = baseline.cross(v2);
+        let d1 = (c1 * c1) / baseline.square_length();
+        let d2 = (c2 * c2) / baseline.square_length();
+
+        let factor = if (c1 * c2) > S::ZERO {
+            S::THREE / S::FOUR
+        } else {
+            S::FOUR / S::NINE
+        };
+
+        let f2 = factor * factor;
+        let threshold = tolerance * tolerance;
+
+        d1 * f2 <= threshold && d2 <= threshold
     }
 
     pub(crate) fn is_a_point(&self, tolerance: S) -> bool {
@@ -323,52 +338,69 @@ impl<S: Scalar> CubicBezierSegment<S> {
         }
     }
 
+    /// Approximate the curve with a single quadratic bézier segment.
+    ///
+    /// This is terrible as a general approximation but works if the cubic
+    /// curve does not have inflection points and is "flat" enough. Typically
+    /// usable after subdividing the curve a few times.
+    pub fn to_quadratic(&self) -> QuadraticBezierSegment<S> {
+        let c1 = (self.ctrl1 * S::THREE - self.from) * S::HALF;
+        let c2 = (self.ctrl2 * S::THREE - self.to) * S::HALF;
+        QuadraticBezierSegment {
+            from: self.from,
+            ctrl: ((c1 + c2) * S::HALF).to_point(),
+            to: self.to,
+        }
+    }
+
+    /// Evaluates an upper bound on the maximum distance between the curve
+    /// and its quadratic approximation obtained using `to_quadratic`.
+    pub fn to_quadratic_error(&self) -> S {
+        // See http://caffeineowl.com/graphics/2d/vectorial/cubic2quad01.html
+        S::sqrt(S::THREE) / S::value(36.0)
+            * ((self.to - self.ctrl2 * S::THREE) + (self.ctrl1 * S::THREE - self.from)).length()
+    }
+
+    /// Returns true if the curve can be safely approximated with a single quadratic bézier
+    /// segment given the provided tolerance threshold.
+    ///
+    /// Equivalent to comparing `to_quadratic_error` with the tolerance threshold, avoiding
+    /// the cost of two square roots.
+    pub fn is_quadratic(&self, tolerance: S) -> bool {
+        S::THREE / S::value(1296.0)
+            * ((self.to - self.ctrl2 * S::THREE) + (self.ctrl1 * S::THREE - self.from))
+                .square_length()
+            <= tolerance * tolerance
+    }
+
+    /// Computes the number of quadratic bézier segments required to approximate this cubic curve
+    /// given a tolerance threshold.
+    ///
+    /// Derived by Raph Levien from section 10.6 of Sedeberg's CAGD notes
+    /// <https://scholarsarchive.byu.edu/cgi/viewcontent.cgi?article=1000&context=facpub#section.10.6>
+    /// and the error metric from the caffein owl blog post <http://caffeineowl.com/graphics/2d/vectorial/cubic2quad01.html>
+    pub fn num_quadratics(&self, tolerance: S) -> u32 {
+        self.num_quadratics_impl(tolerance).to_u32().unwrap_or(1)
+    }
+
+    fn num_quadratics_impl(&self, tolerance: S) -> S {
+        debug_assert!(tolerance > S::ZERO);
+
+        let x = self.from.x - S::THREE * self.ctrl1.x + S::THREE * self.ctrl2.x - self.to.x;
+        let y = self.from.y - S::THREE * self.ctrl1.y + S::THREE * self.ctrl2.y - self.to.y;
+
+        let err = x * x + y * y;
+
+        (err / (S::value(432.0) * tolerance * tolerance))
+            .powf(S::ONE / S::SIX)
+            .ceil()
+            .max(S::ONE)
+    }
+
     /// Returns the flattened representation of the curve as an iterator, starting *after* the
     /// current point.
     pub fn flattened(&self, tolerance: S) -> Flattened<S> {
         Flattened::new(self, tolerance)
-    }
-
-    /// Invokes a callback between each monotonic part of the segment.
-    fn for_each_monotonic_t<F>(&self, mut cb: F)
-    where
-        F: FnMut(S),
-    {
-        let mut x_extrema: ArrayVec<S, 3> = ArrayVec::new();
-        self.for_each_local_x_extremum_t(&mut |t| x_extrema.push(t));
-
-        let mut y_extrema: ArrayVec<S, 3> = ArrayVec::new();
-        self.for_each_local_y_extremum_t(&mut |t| y_extrema.push(t));
-
-        let mut it_x = x_extrema.iter().cloned();
-        let mut it_y = y_extrema.iter().cloned();
-        let mut tx = it_x.next();
-        let mut ty = it_y.next();
-        loop {
-            let next = match (tx, ty) {
-                (Some(a), Some(b)) => {
-                    if a < b {
-                        tx = it_x.next();
-                        a
-                    } else {
-                        ty = it_y.next();
-                        b
-                    }
-                }
-                (Some(a), None) => {
-                    tx = it_x.next();
-                    a
-                }
-                (None, Some(b)) => {
-                    ty = it_y.next();
-                    b
-                }
-                (None, None) => return,
-            };
-            if next > S::ZERO && next < S::ONE {
-                cb(next);
-            }
-        }
     }
 
     /// Invokes a callback for each monotonic part of the segment.
@@ -376,12 +408,20 @@ impl<S: Scalar> CubicBezierSegment<S> {
     where
         F: FnMut(Range<S>),
     {
+        let mut extrema: ArrayVec<S, 4> = ArrayVec::new();
+        self.for_each_local_x_extremum_t(&mut |t| extrema.push(t));
+        self.for_each_local_y_extremum_t(&mut |t| extrema.push(t));
+        extrema.sort_unstable_by(|a, b| a.partial_cmp(b).unwrap());
+
         let mut t0 = S::ZERO;
-        self.for_each_monotonic_t(|t| {
-            cb(t0..t);
-            t0 = t;
-        });
-        cb(t0..S::ONE);
+        for &t in &extrema {
+            if t != t0 {
+                cb(t0 .. t);
+                t0 = t;
+            }
+        }
+
+        cb(t0 .. S::ONE);
     }
 
     /// Invokes a callback for each monotonic part of the segment.
@@ -473,7 +513,7 @@ impl<S: Scalar> CubicBezierSegment<S> {
     where
         F: FnMut(&QuadraticBezierSegment<S>),
     {
-        cubic_to_quadratics(self, tolerance, cb);
+        self.for_each_quadratic_bezier_with_t(tolerance, &mut |quad, _range| cb(quad));
     }
 
     /// Approximates the cubic bézier curve with sequence of quadratic ones,
@@ -482,38 +522,81 @@ impl<S: Scalar> CubicBezierSegment<S> {
     where
         F: FnMut(&QuadraticBezierSegment<S>, Range<S>),
     {
-        cubic_to_quadratics_with_t(self, tolerance, cb);
-    }
+        debug_assert!(tolerance >= S::EPSILON * S::EPSILON);
 
-    /// Approximates the cubic bézier curve with sequence of monotonic quadratic
-    /// ones, invoking a callback at each step.
-    pub fn for_each_monotonic_quadratic<F>(&self, tolerance: S, cb: &mut F)
-    where
-        F: FnMut(&QuadraticBezierSegment<S>),
-    {
-        cubic_to_monotonic_quadratics(self, tolerance, cb);
+        let num_quadratics = self.num_quadratics_impl(tolerance);
+        let step = S::ONE / num_quadratics;
+        let n = num_quadratics.to_u32().unwrap_or(1);
+        let mut t0 = S::ZERO;
+        for _ in 0..(n - 1) {
+            let t1 = t0 + step;
+
+            let quad = self.split_range(t0..t1).to_quadratic();
+            cb(&quad, t0..t1);
+
+            t0 = t1;
+        }
+
+        // Do the last step manually to make sure we finish at t = 1.0 exactly.
+        let quad = self.split_range(t0..S::ONE).to_quadratic();
+        cb(&quad, t0..S::ONE)
     }
 
     /// Compute a flattened approximation of the curve, invoking a callback at
     /// each step.
     pub fn for_each_flattened<F: FnMut(Point<S>)>(&self, tolerance: S, callback: &mut F) {
-        flatten_cubic_bezier_with_t(self, tolerance, &mut |point, _| callback(point));
+        debug_assert!(tolerance >= S::EPSILON * S::EPSILON);
+        let quadratics_tolerance = tolerance * S::value(0.4);
+        let flattening_tolerance = tolerance * S::value(0.8);
 
-        callback(self.to);
+        self.for_each_quadratic_bezier(quadratics_tolerance, &mut |quad| {
+            quad.for_each_flattened(flattening_tolerance, &mut |point| {
+                callback(point);
+            });
+        });
     }
 
     /// Compute a flattened approximation of the curve, invoking a callback at
     /// each step, including the final endpoint.
     pub fn for_each_flattened_with_t<F: FnMut(Point<S>, S)>(&self, tolerance: S, callback: &mut F) {
-        flatten_cubic_bezier_with_t(self, tolerance, callback);
+        debug_assert!(tolerance >= S::EPSILON * S::EPSILON);
+        let quadratics_tolerance = tolerance * S::value(0.4);
+        let flattening_tolerance = tolerance * S::value(0.8);
 
-        callback(self.to, S::ONE);
+        self.for_each_quadratic_bezier_with_t(quadratics_tolerance, &mut |quad, range| {
+            let range_len = range.end - range.start;
+            quad.for_each_flattened_with_t(flattening_tolerance, &mut |point, t_sub| {
+                let t = t_sub * range_len + range.start;
+                callback(point, t);
+            });
+        });
     }
 
     /// Compute a flattened approximation of the curve, invoking a callback at
     /// each step, excluding the final endpoint.
     pub fn for_each_flattened_with_t_intermediate<F: FnMut(Point<S>, S)>(&self, tolerance: S, callback: &mut F) {
-        flatten_cubic_bezier_with_t(self, tolerance, callback);
+        debug_assert!(tolerance >= S::EPSILON * S::EPSILON);
+        let quadratics_tolerance = tolerance * S::value(0.4);
+        let flattening_tolerance = tolerance * S::value(0.8);
+
+        let num_quadratics = self.num_quadratics_impl(quadratics_tolerance);
+        let step = S::ONE / num_quadratics;
+        let n = num_quadratics.to_u32().unwrap_or(1);
+        let mut t0 = S::ZERO;
+        for _ in 0..(n - 1) {
+            let t1 = t0 +  step;
+
+            let quad = self.split_range(t0 .. t1).to_quadratic();
+            quad.for_each_flattened_with_t(flattening_tolerance, &mut |point, t_sub| {
+                let t = t_sub * step + t0;
+                callback(point, t);
+            });
+
+            t0 = t1;
+        }
+
+        let last_quad = self.split_range(t0 .. S::ONE).to_quadratic();
+        last_quad.for_each_flattened_with_t_intermediate(flattening_tolerance, callback);
     }
 
     /// Compute the length of the segment using a flattened approximation.
@@ -528,11 +611,87 @@ impl<S: Scalar> CubicBezierSegment<S> {
         len
     }
 
+    /// Invokes a callback at each inflection point if any.
     pub fn for_each_inflection_t<F>(&self, cb: &mut F)
     where
         F: FnMut(S),
     {
-        find_cubic_bezier_inflection_points(self, cb);
+        // Find inflection points.
+        // See www.faculty.idc.ac.il/arik/quality/appendixa.html for an explanation
+        // of this approach.
+        let pa = self.ctrl1 - self.from;
+        let pb =
+            self.ctrl2.to_vector() - (self.ctrl1.to_vector() * S::TWO) + self.from.to_vector();
+        let pc = self.to.to_vector() - (self.ctrl2.to_vector() * S::THREE)
+            + (self.ctrl1.to_vector() * S::THREE)
+            - self.from.to_vector();
+
+        let a = pb.cross(pc);
+        let b = pa.cross(pc);
+        let c = pa.cross(pb);
+
+        if S::abs(a) < S::EPSILON {
+            // Not a quadratic equation.
+            if S::abs(b) < S::EPSILON {
+                // Instead of a linear acceleration change we have a constant
+                // acceleration change. This means the equation has no solution
+                // and there are no inflection points, unless the constant is 0.
+                // In that case the curve is a straight line, essentially that means
+                // the easiest way to deal with is is by saying there's an inflection
+                // point at t == 0. The inflection point approximation range found will
+                // automatically extend into infinity.
+                if S::abs(c) < S::EPSILON {
+                    cb(S::ZERO);
+                }
+            } else {
+                let t = -c / b;
+                if in_range(t) {
+                    cb(t);
+                }
+            }
+
+            return;
+        }
+
+        fn in_range<S: Scalar>(t: S) -> bool {
+            t >= S::ZERO && t < S::ONE
+        }
+
+        let discriminant = b * b - S::FOUR * a * c;
+
+        if discriminant < S::ZERO {
+            return;
+        }
+
+        if discriminant < S::EPSILON {
+            let t = -b / (S::TWO * a);
+
+            if in_range(t) {
+                cb(t);
+            }
+
+            return;
+        }
+
+        // This code is derived from https://www2.units.it/ipl/students_area/imm2/files/Numerical_Recipes.pdf page 184.
+        // Computing the roots this way avoids precision issues when a, c or both are small.
+        let discriminant_sqrt = S::sqrt(discriminant);
+        let sign_b = if b >= S::ZERO { S::ONE } else { -S::ONE };
+        let q = -S::HALF * (b + sign_b * discriminant_sqrt);
+        let mut first_inflection = q / a;
+        let mut second_inflection = c / q;
+
+        if first_inflection > second_inflection {
+            std::mem::swap(&mut first_inflection, &mut second_inflection);
+        }
+
+        if in_range(first_inflection) {
+            cb(first_inflection);
+        }
+
+        if in_range(second_inflection) {
+            cb(second_inflection);
+        }
     }
 
     /// Return local x extrema or None if this curve is monotonic.
@@ -1030,6 +1189,284 @@ impl<S: Scalar> BoundingBox for CubicBezierSegment<S> {
     }
 }
 
+use crate::quadratic_bezier::FlattenedT as FlattenedQuadraticSegment;
+
+pub struct Flattened<S: Scalar> {
+    curve: CubicBezierSegment<S>,
+    current_curve: FlattenedQuadraticSegment<S>,
+    remaining_sub_curves: i32,
+    tolerance: S,
+    range_step: S,
+    range_start: S,
+}
+
+impl<S: Scalar> Flattened<S> {
+    pub(crate) fn new(curve: &CubicBezierSegment<S>, tolerance: S) -> Self {
+        debug_assert!(tolerance >= S::EPSILON * S::EPSILON);
+
+        let quadratics_tolerance = tolerance * S::value(0.4);
+        let flattening_tolerance = tolerance * S::value(0.8);
+
+        let num_quadratics = curve.num_quadratics_impl(quadratics_tolerance);
+
+        let range_step = S::ONE / num_quadratics;
+
+        let quadratic = curve.split_range(S::ZERO..range_step).to_quadratic();
+        let current_curve = FlattenedQuadraticSegment::new(&quadratic, flattening_tolerance);
+
+        Flattened {
+            curve: *curve,
+            current_curve,
+            remaining_sub_curves: num_quadratics.to_i32().unwrap() - 1,
+            tolerance: flattening_tolerance,
+            range_start: S::ZERO,
+            range_step,
+        }
+    }
+}
+
+impl<S: Scalar> Iterator for Flattened<S> {
+    type Item = Point<S>;
+
+    fn next(&mut self) -> Option<Point<S>> {
+        if let Some(t_inner) = self.current_curve.next() {
+            let t = self.range_start + t_inner * self.range_step;
+            return Some(self.curve.sample(t));
+        }
+
+        if self.remaining_sub_curves <= 0 {
+            return None;
+        }
+
+        self.range_start += self.range_step;
+        let t0 = self.range_start;
+        let t1 = self.range_start + self.range_step;
+        self.remaining_sub_curves -= 1;
+
+        let quadratic = self.curve.split_range(t0..t1).to_quadratic();
+        self.current_curve = FlattenedQuadraticSegment::new(&quadratic, self.tolerance);
+
+        let t_inner = self.current_curve.next().unwrap_or(S::ONE);
+        let t = t0 + t_inner * self.range_step;
+
+        Some(self.curve.sample(t))
+    }
+
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        (
+            self.remaining_sub_curves as usize * self.current_curve.size_hint().0,
+            None,
+        )
+    }
+}
+
+#[cfg(test)]
+fn print_arrays(a: &[Point<f32>], b: &[Point<f32>]) {
+    println!("left:  {:?}", a);
+    println!("right: {:?}", b);
+}
+
+#[cfg(test)]
+fn assert_approx_eq(a: &[Point<f32>], b: &[Point<f32>]) {
+    if a.len() != b.len() {
+        print_arrays(a, b);
+        panic!("Lengths differ ({} != {})", a.len(), b.len());
+    }
+    for i in 0..a.len() {
+        let threshold = 0.029;
+        let dx = f32::abs(a[i].x - b[i].x);
+        let dy = f32::abs(a[i].y - b[i].y);
+        if dx > threshold || dy > threshold {
+            print_arrays(a, b);
+            println!("diff = {:?} {:?}", dx, dy);
+            panic!("The arrays are not equal");
+        }
+    }
+}
+
+#[test]
+fn test_iterator_builder_1() {
+    let tolerance = 0.01;
+    let c1 = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(1.0, 0.0),
+        ctrl2: Point::new(1.0, 1.0),
+        to: Point::new(0.0, 1.0),
+    };
+    let iter_points: Vec<Point<f32>> = c1.flattened(tolerance).collect();
+    let mut builder_points = Vec::new();
+    c1.for_each_flattened(tolerance, &mut |p| {
+        builder_points.push(p);
+    });
+
+    assert!(iter_points.len() > 2);
+    assert_approx_eq(&iter_points[..], &builder_points[..]);
+}
+
+#[test]
+fn test_iterator_builder_2() {
+    let tolerance = 0.01;
+    let c1 = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(1.0, 0.0),
+        ctrl2: Point::new(0.0, 1.0),
+        to: Point::new(1.0, 1.0),
+    };
+    let iter_points: Vec<Point<f32>> = c1.flattened(tolerance).collect();
+    let mut builder_points = Vec::new();
+    c1.for_each_flattened(tolerance, &mut |p| {
+        builder_points.push(p);
+    });
+
+    assert!(iter_points.len() > 2);
+    assert_approx_eq(&iter_points[..], &builder_points[..]);
+}
+
+#[test]
+fn test_iterator_builder_3() {
+    let tolerance = 0.01;
+    let c1 = CubicBezierSegment {
+        from: Point::new(141.0, 135.0),
+        ctrl1: Point::new(141.0, 130.0),
+        ctrl2: Point::new(140.0, 130.0),
+        to: Point::new(131.0, 130.0),
+    };
+    let iter_points: Vec<Point<f32>> = c1.flattened(tolerance).collect();
+    let mut builder_points = Vec::new();
+    c1.for_each_flattened(tolerance, &mut |p| {
+        builder_points.push(p);
+    });
+
+    assert!(iter_points.len() > 2);
+    assert_approx_eq(&iter_points[..], &builder_points[..]);
+}
+
+#[test]
+fn test_issue_19() {
+    let tolerance = 0.15;
+    let c1 = CubicBezierSegment {
+        from: Point::new(11.71726, 9.07143),
+        ctrl1: Point::new(1.889879, 13.22917),
+        ctrl2: Point::new(18.142855, 19.27679),
+        to: Point::new(18.142855, 19.27679),
+    };
+    let iter_points: Vec<Point<f32>> = c1.flattened(tolerance).collect();
+    let mut builder_points = Vec::new();
+    c1.for_each_flattened(tolerance, &mut |p| {
+        builder_points.push(p);
+    });
+
+    assert_approx_eq(&iter_points[..], &builder_points[..]);
+
+    assert!(iter_points.len() > 1);
+}
+
+#[test]
+fn test_issue_194() {
+    let segment = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(0.0, 0.0),
+        ctrl2: Point::new(50.0, 70.0),
+        to: Point::new(100.0, 100.0),
+    };
+
+    let mut points = Vec::new();
+    segment.for_each_flattened(0.1, &mut |p| {
+        points.push(p);
+    });
+
+    assert!(points.len() > 2);
+}
+
+#[test]
+fn flatten_with_t() {
+    let segment = CubicBezierSegment {
+        from: Point::new(0.0f32, 0.0),
+        ctrl1: Point::new(0.0, 0.0),
+        ctrl2: Point::new(50.0, 70.0),
+        to: Point::new(100.0, 100.0),
+    };
+
+    for tolerance in &[0.1, 0.01, 0.001, 0.0001] {
+        let tolerance = *tolerance;
+
+        let mut a = Vec::new();
+        segment.for_each_flattened(tolerance, &mut |p| {
+            a.push(p);
+        });
+
+        let mut b = Vec::new();
+        let mut ts = Vec::new();
+        segment.for_each_flattened_with_t(tolerance, &mut |p, t| {
+            b.push(p);
+            ts.push(t);
+        });
+
+        assert_eq!(a, b);
+
+        for i in 0..b.len() {
+            let sampled = segment.sample(ts[i]);
+            let point = b[i];
+            let dist = (sampled - point).length();
+            assert!(dist <= tolerance);
+        }
+    }
+}
+
+#[test]
+fn test_flatten_end() {
+    let segment = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(100.0, 0.0),
+        ctrl2: Point::new(100.0, 100.0),
+        to: Point::new(100.0, 200.0),
+    };
+
+    let mut last = segment.from;
+    segment.for_each_flattened(0.0001, &mut |p| {
+        last = p;
+    });
+
+    assert_eq!(last, segment.to);
+}
+
+#[test]
+fn test_flatten_point() {
+    let segment = CubicBezierSegment {
+        from: Point::new(0.0, 0.0),
+        ctrl1: Point::new(0.0, 0.0),
+        ctrl2: Point::new(0.0, 0.0),
+        to: Point::new(0.0, 0.0),
+    };
+
+    let mut last = segment.from;
+    segment.for_each_flattened(0.0001, &mut |p| {
+        last = p;
+    });
+
+    assert_eq!(last, segment.to);
+}
+
+#[test]
+fn issue_652() {
+    use crate::point;
+
+    let curve = CubicBezierSegment {
+        from: point(-1061.0, -3327.0),
+        ctrl1: point(-1061.0, -3177.0),
+        ctrl2: point(-1061.0, -3477.0),
+        to: point(-1061.0, -3327.0),
+    };
+
+    for _ in curve.flattened(1.0) {}
+    for _ in curve.flattened(0.1) {}
+    for _ in curve.flattened(0.01) {}
+
+    curve.for_each_flattened(1.0, &mut |_| {});
+    curve.for_each_flattened(0.1, &mut |_| {});
+    curve.for_each_flattened(0.01, &mut |_| {});
+}
+
 #[test]
 fn fast_bounding_box_for_cubic_bezier_segment() {
     let a = CubicBezierSegment {
@@ -1452,4 +1889,44 @@ fn cubic_line_intersection_on_endpoint() {
     let intersections = c1.cubic_intersections_t(&c2);
 
     assert!(intersections.is_empty());
+}
+
+
+#[test]
+fn test_cubic_to_quadratics() {
+    use euclid::approxeq::ApproxEq;
+
+    let quadratic = QuadraticBezierSegment {
+        from: point(1.0, 2.0),
+        ctrl: point(10.0, 5.0),
+        to: point(0.0, 1.0),
+    };
+
+    let mut count = 0;
+    assert_eq!(quadratic.to_cubic().num_quadratics(0.0001), 1);
+    quadratic.to_cubic().for_each_quadratic_bezier(0.0001, &mut |c| {
+        assert!(count == 0);
+        assert!(c.from.approx_eq(&quadratic.from));
+        assert!(c.ctrl.approx_eq(&quadratic.ctrl));
+        assert!(c.to.approx_eq(&quadratic.to));
+        count += 1;
+    });
+
+    let cubic = CubicBezierSegment {
+        from: point(1.0f32, 1.0),
+        ctrl1: point(10.0, 2.0),
+        ctrl2: point(1.0, 3.0),
+        to: point(10.0, 4.0),
+    };
+
+    let mut prev = cubic.from;
+    let mut count = 0;
+    cubic.for_each_quadratic_bezier(0.01, &mut |c| {
+        assert!(c.from.approx_eq(&prev));
+        prev = c.to;
+        count += 1;
+    });
+    assert!(prev.approx_eq(&cubic.to));
+    assert!(count < 10);
+    assert!(count > 4);
 }
